Fix a data type error in Document

+++ search-server/main.cpp
@@ -46,7 +46,7 @@ vector<string> SplitIntoWords(const string& text) {
 
 struct Document {
     int id;
-    int relevance;
+    double relevance;
 };
 
 class SearchServer {
